@@ -5,7 +5,7 @@ requires = ["setuptools", "wheel", "versioningit"]
 [project]
 name = "adaptive_scheduler"
 description = "Run many `adaptive.Learner`s on many cores (>10k) using `mpi4py.futures`, `ipyparallel`, `dask-mpi`, or `process-pool`."
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 dynamic = ["version"]
 maintainers = [{ name = "Bas Nijholt", email = "bas@nijho.lt" }]
 license = { text = "BSD-3" }
@@ -17,10 +17,9 @@ classifiers = [
     "Operating System :: OS Independent",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
     "Topic :: Scientific/Engineering",
     "Topic :: System :: Distributed Computing",
 ]
@@ -111,7 +110,7 @@ exclude_lines = [
 
 [tool.ruff]
 line-length = 88
-target-version = "py38"
+target-version = "py310"
 
 [tool.ruff.lint]
 select = ["ALL"]
@@ -143,4 +142,4 @@ ignore = [
 max-complexity = 18
 
 [tool.mypy]
-python_version = "3.8"
+python_version = "3.10"
